Fixing a few unsafe typecasts
#rb jimmy.andrews nathan.mitchell
#preflight 62b213171b2d3b6cafd334e5

[CL 20763388 by tyson brochu in ue5-main branch]

+++ Engine/Plugins/Runtime/GeometryProcessing/Source/GeometryAlgorithms/Private/XAtlasWrapper.cpp
@@ -88,8 +88,8 @@ namespace XAtlasWrapper
 		UVIndexBuffer.Empty();
 		VertexRemapArray.Empty();
 
-		const float W = Atlas->width;
-		const float H = Atlas->height;
+		const double W = Atlas->width;
+		const double H = Atlas->height;
 
 		for (uint32 OutMeshID = 0; OutMeshID < Atlas->meshCount; ++OutMeshID)
 		{
